Added procedural texture

+++ tasks/local_shadertoy2/Alignstd430.hpp
@@ -183,7 +183,7 @@ template <glsl_mat T>
 size_t memcpy_std430(char* buf, size_t offset, T& val)
 {
   typename T::row_type tmp[T::length()];
-  for (size_t i = 0; i < T::length(); ++i)
+  for (size_t i = 0; i < static_cast<size_t>(T::length()); ++i)
   {
     tmp[i] = val[i];
   }
